[SYSTEMDS-3644] Compressed transform encode
This commit primarily contains optimizations to compressed and uncompressed transform encode. We observe 3x improved speed from various uncompressed or compressed transformations. The primary source of speedup is a custom HashMap implementation that leverages primitive int values for the recode map.

The commit also contains minor optimizations to reading FrameBlock CSV files and cleanup to the logic in CommonThreadPool to not allocate new threads if only 1 thread is requested.

Closes #2171

+++ src/main/java/org/apache/sysds/runtime/instructions/spark/MultiReturnParameterizedBuiltinSPInstruction.java
@@ -325,7 +325,7 @@ public class MultiReturnParameterizedBuiltinSPInstruction extends ComputationSPI
 			Iterator<Object> iter = arg0._2().iterator();
 			ArrayList<String> ret = new ArrayList<>();
 
-			long rowID = 1;
+			int rowID = 1;
 			StringBuilder sb = new StringBuilder();
 
 			// handle recode maps
@@ -371,7 +371,7 @@ public class MultiReturnParameterizedBuiltinSPInstruction extends ComputationSPI
 			else {
 				throw new DMLRuntimeException("Unsupported metadata output for encoder: \n" + _encoder);
 			}
-			_accMax.add(rowID - 1);
+			_accMax.add(rowID - 1L);
 
 			return ret.iterator();
 		}
